WebApp/server/requirements.txt
controlnet_aux==0.0.7
Flask==3.0.0
Flask_SocketIO==5.3.6
matplotlib==3.7.2
numpy==1.26.3
opencv_contrib_python==4.8.0.74
opencv_python==4.8.1.78
Pillow==10.2.0
